Implement alu decoder
This implements a subset of all RISCV32I arithmetic/logical instructions
as there are 10 of them, whereas the project brief specifies ALUControl
to be 3 bits.

Look at list of R-type instructions at end of lecture 6 for all possible
ALU functions (I think). The ones not implemented are:
   shift left logic
   set less than unsigned
   xor
   shift right logical
   shift right arithmetic

The remaining arithmetic/logical operations are used by arithmetic/logical
instructions, as well as these instructions:
    lw, sw (to add register addresses together)
    beq    (to subtract operands; if the result is 0, this asserts the zero
            flag, which is then used to branch)