Update PC
Updated PC bit width to full 32-bits, to be
in line with the changes made to data_mem
and instr_mem (to make their addresses 32 bit
also).

Integrated the three muxes for trigger, rst and
pcsrc into a single mux. This greatly simplifies
the code, as the desired behaviour is described
behaviourly instead of structurally (see lecture
2 slide 5).