Implement fix for JALR
Fixed bug that occurs if JALR is executed when rd != x0.

Previously, the output of the result mux was fed both into
pc_mux as well as reg_file. The intention was that:
     --> reg_file  would receive pcplus4   to perform rd = PC + 4
         (which is input to the result mux)
     --> pc_mux    would receive aluresult to perform PC = rs1 + SignExt(imm)
         (which is *also* input to the result mux).

These are two different inputs to result mux that are
intended to go to two different destinations (reg_file
and pc_mux). But the result mux only has a single output,
so only one of these can be done with result_mux.

Incidentally, the first output is correct as that is also
used for JAL/BEQ. Hence, we must change how pc_mux receives
aluresult. This is simple: instead of feeding it out of the
result mux, we can take it directly from the alu.

This solves the bug, as can be verified with the jalr_test
program.