Update ALU to handle all possible operations
The ALU can now carry out all of the following:
   add     sub
   slt     xor
   srl     sra
   or      and
   slt     sltu

Depending on what is input, the immediate version
of these instructions is also available.

To do this, alucontrol had to be made 4 bits from
its previous 3 (to fit 10 possible ALU actions).

The purpose of this addition was primarily to use
slt and sltu for blt, bge, bltu, bgeu. The other
instructions were an easy addition after that.

Run the alu testbench to see each instruction working.
This only tests alu, and not top_alu, as testing top_alu
is presently unnecessary.