Parity update with base_dev
Update the piepleined CPU's code to resemble the latest version of the base CPU. Work on pipelining began before the base
cpu was finished, which causes the two codebases to diverge more than they should, making working across branches difficult
(e.g. if a bug is solved in base but not in piepelined, but you don't know that).

Many changes have been made, but ultimately the code is more similar to the base cpu except where it must differ to implement
pipelining.