Connect data mem and control unit
This is done via a new 3-bit control signal called memop. This
signal is equivalent to funct3. It has been made an output from
the control unit rather than being connected to data memory directly
to make pipelining easier (as this allows memop to propagate through
different stages with the rest of the relevant control signals).