Further testing with pdf program
Created pdf_expected_output.cpp, which should do the same thing
as the pdf reference program, but in cpp. Mainly used as a sanity
check, but the the printed values can be copied in to a spreadsheet,
plotted, and compared against the Vbuddy plot. Visually, they look
the same.

Adjusted cpu_vbuddy_tb.cpp to only plot a value once every 3 cycles,
to account for the pdf program taking 3 cycles in its display loop.

To make it even more seamless, every 256 cycles you could stop plotting
for 5 cycles to account for the cycles taken to exit and re-enter
_loop3.